requirement.txt.txt
asgiref==3.4.1
Django==3.2.6
django-cleanup==5.2.0
django-taggit==1.5.1
easy-thumbnails==2.7.1
Markdown==3.3.4
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
